feat : 커스텀 응답 DTO의 응답 코드 200으로 수정 및 data.sql 추가

+++ src/main/java/com/pjb/topicboard/domain/comment/service/CommentService.java
@@ -63,6 +63,10 @@ public class CommentService {
                 () -> new CustomCommonException(ErrorEnum.COMMENT_NOT_FOUND)
         );
 
+        if(!comment.getAuthor().getId().equals(customUserDetails.getUser().getId())){
+            throw new CustomCommonException(ErrorEnum.FORBIDDEN_USER);
+        }
+
         commentRepository.delete(comment);
     }
